use the script element to contain the markdown
A script element contains CDATA. That means that it
can contain whatever characters including HTML markup and
will not be messed with by the browser. That means that we
can use ampersands ('&') throughout the markdown code and
it will not be replaced by the browser with &amp;

That way showdown.js' makeHTML can do it's work properly and
translate the markdown into html.

* see https://github.com/preek/p_slides/issues/1